+++ Changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `.release` class method for many `GDAL` types.
+- Added `GDAL::Dataset.copy_whole_raster`.
 - Added exception messages for `OGR` methods that raised using
   `OGR::ErrorHandling`.